src/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
Pillow>=7.0.0
numpy>=1.18.0
matplotlib>=3.2.0
scikit-learn>=0.23.0
seaborn>=0.11.0
tqdm>=4.50.0